chore: added tests for uint types

+++ model/ledger/signer_list.go
@@ -19,8 +19,8 @@ type SignerList struct {
 	PreviousTxnLgrSeq uint32
 	OwnerNode         string
 	SignerEntries     []SignerEntryWrapper
-	SignerListID      uint64
-	SignerQuorum      uint64
+	SignerListID      uint32
+	SignerQuorum      uint32
 }
 
 type SignerEntryWrapper struct {
@@ -29,7 +29,7 @@ type SignerEntryWrapper struct {
 
 type SignerEntry struct {
 	Account       types.Address
-	SignerWeight  uint64
+	SignerWeight  uint16
 	WalletLocator types.Hash256 `json:",omitempty"`
 }
